added DINA submodule
