+++ src/hisp/dina/__init__.py
@@ -1,0 +1,1 @@
+from .dina import RISP_data, heat, get_particle_flux
